ContainerDevice: add AttachError to handle device attach error
Got "docker: Error response from daemon: OCI runtime create failed:
QMP command failed: unknown." when "docker run --privileged" with kata.
The cause is a device got a error when attach to sandbox.

This error doesn't affect the execution of kata.  So just output the
error and record it in AttachError let kataAgent ignore it.

Fixes #829

Signed-off-by: Hui Zhu <teawater@hyper.sh>